+++ include/matchable.h
@@ -3,14 +3,16 @@
 
 #include <string>
 
+#include "3rd-party/optional.hpp"
+
 namespace newsboat {
 
 class Matchable {
 public:
 	Matchable() = default;
 	virtual ~Matchable() = default;
-	virtual bool has_attribute(const std::string& attribname) = 0;
-	virtual std::string get_attribute(const std::string& attribname) = 0;
+	virtual nonstd::optional<std::string> attribute_value(const std::string& attr) =
+		0;
 };
 
 } // namespace newsboat